refactor: extract useOperationForm hook and integrate into OperationModal
- Create useOperationForm hook with form state, validation, save/delete logic
- Integrate multi-currency transfer logic into useOperationForm
- Use useOperationPicker for category navigation
- Remove 356 lines of duplicate code from OperationModal
- Reduce OperationModal from 1,074 to 718 lines (33% reduction)
