+++ app/modals/OperationModal.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo, useCallback } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import {
   View,
@@ -22,15 +22,15 @@ import { useDialog } from '../contexts/DialogContext';
 import { useOperations } from '../contexts/OperationsContext';
 import { useAccounts } from '../contexts/AccountsContext';
 import { useCategories } from '../contexts/CategoriesContext';
-import { getLastAccessedAccount, setLastAccessedAccount } from '../services/LastAccount';
-import Calculator from '../components/Calculator';
+import { setLastAccessedAccount } from '../services/LastAccount';
 import OperationFormFields from '../components/operations/OperationFormFields';
 import * as Currency from '../services/currency';
 import { formatDate } from '../services/BalanceHistoryDB';
 import { SPACING, BORDER_RADIUS } from '../styles/designTokens';
 import currencies from '../../assets/currencies.json';
 import { hasOperation, evaluateExpression } from '../utils/calculatorUtils';
-import { getCategoryDisplayName } from '../utils/categoryUtils';
+import useOperationForm from '../hooks/useOperationForm';
+import useOperationPicker from '../hooks/useOperationPicker';
 
 /**
  * OperationModal Component
@@ -66,31 +66,51 @@ export default function OperationModal({ visible, onClose, operation, isNew, onD
   const { visibleAccounts: accounts } = useAccounts();
   const { categories } = useCategories();
 
-  const [values, setValues] = useState({
-    type: 'expense',
-    amount: '',
-    accountId: '',
-    categoryId: '',
-    date: formatDate(new Date()),
-    toAccountId: '',
-    exchangeRate: '',
-    destinationAmount: '',
+  // Operation form hook (includes multi-currency logic)
+  const {
+    values,
+    setValues,
+    errors,
+    showDatePicker,
+    setShowDatePicker,
+    lastEditedField,
+    setLastEditedField,
+    isShadowOperation,
+    canDeleteShadowOperation,
+    filteredCategories,
+    sourceAccount,
+    destinationAccount,
+    isMultiCurrencyTransfer,
+    handleSave,
+    handleClose,
+    handleDelete,
+    getAccountName,
+    getCategoryName,
+    formatDateForDisplay,
+  } = useOperationForm({
+    visible,
+    operation,
+    isNew,
+    accounts,
+    categories,
+    t,
+    addOperation,
+    updateOperation,
+    validateOperation,
+    showDialog,
+    onClose,
+    onDelete,
   });
-  const [errors, setErrors] = useState({});
-  const [showDatePicker, setShowDatePicker] = useState(false);
-  const [pickerState, setPickerState] = useState({
-    visible: false,
-    type: null,
-    data: [],
-    allCategories: [],
-  });
-  // Category navigation state (for hierarchical folder navigation)
-  const [categoryNavigation, setCategoryNavigation] = useState({
-    currentFolderId: null,
-    breadcrumb: [],
-  });
-  // Track which field was last edited to determine calculation direction
-  const [lastEditedField, setLastEditedField] = useState(null);
+
+  // Operation picker hook for category navigation
+  const {
+    pickerState,
+    categoryNavigation,
+    openPicker,
+    closePicker,
+    navigateIntoFolder,
+    navigateBack,
+  } = useOperationPicker(t);
 
   // Memoize calculator amount change handler for performance
   const handleAmountChange = useCallback((text) => {
@@ -98,382 +118,7 @@ export default function OperationModal({ visible, onClose, operation, isNew, onD
       setValues(v => ({ ...v, amount: text }));
       setLastEditedField('amount');
     }
-  }, [isShadowOperation]);
-
-  useEffect(() => {
-    // Only run when modal becomes visible
-    if (!visible) return;
-
-    let cancelled = false;
-    async function setDefaultAccount() {
-      if (operation && !isNew) {
-        // Normalize values when editing an existing operation
-        setValues({
-          type: operation.type || 'expense',
-          amount: String(operation.amount || ''),
-          accountId: operation.accountId || accounts[0]?.id || '',
-          categoryId: operation.categoryId || '',
-          description: operation.description || '',
-          date: operation.date || formatDate(new Date()),
-          toAccountId: operation.toAccountId || '',
-          exchangeRate: String(operation.exchangeRate || ''),
-          destinationAmount: String(operation.destinationAmount || ''),
-        });
-        // Reset last edited field when opening modal with existing operation
-        setLastEditedField(null);
-      } else if (isNew) {
-        let defaultAccountId = '';
-        if (accounts.length === 1) {
-          defaultAccountId = accounts[0].id;
-        } else if (accounts.length > 1) {
-          // Try to get last accessed account
-          const lastId = await getLastAccessedAccount();
-          if (lastId && accounts.some(acc => acc.id === lastId)) {
-            defaultAccountId = lastId;
-          } else {
-            // Fallback: lowest ID (lexicographically)
-            defaultAccountId = accounts.slice().sort((a, b) => (a.id < b.id ? -1 : 1))[0].id;
-          }
-        }
-        if (!cancelled) {
-          setValues({
-            type: 'expense',
-            amount: '',
-            accountId: defaultAccountId,
-            categoryId: '',
-            description: '',
-            date: formatDate(new Date()),
-            toAccountId: '',
-            exchangeRate: '',
-            destinationAmount: '',
-          });
-          // Reset last edited field when opening modal for new operation
-          setLastEditedField(null);
-        }
-      }
-      if (!cancelled) setErrors({});
-    }
-    setDefaultAccount();
-    return () => { cancelled = true; };
-  }, [operation, isNew, visible, accounts]);
-
-  const validateFields = useCallback((valuesToValidate = null) => {
-    const vals = valuesToValidate || values;
-    const newErrors = {};
-
-    if (!vals.type) {
-      newErrors.type = t('operation_type_required');
-    }
-
-    if (!vals.amount || isNaN(parseFloat(vals.amount)) || parseFloat(vals.amount) <= 0) {
-      newErrors.amount = t('valid_amount_required');
-    }
-
-    if (!vals.accountId) {
-      newErrors.accountId = t('account_required');
-    }
-
-    if (vals.type === 'transfer') {
-      if (!vals.toAccountId) {
-        newErrors.toAccountId = t('destination_account_required');
-      } else if (vals.accountId === vals.toAccountId) {
-        newErrors.toAccountId = t('accounts_must_be_different');
-      }
-    } else if (!vals.categoryId) {
-      newErrors.categoryId = t('category_required');
-    }
-
-    if (!vals.date) {
-      newErrors.date = t('date_required');
-    }
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  }, [values, t]);
-
-  const handleSave = useCallback(async () => {
-    // Automatically evaluate any pending math operation before saving
-    let finalAmount = values.amount;
-    if (hasOperation(values.amount)) {
-      const evaluated = evaluateExpression(values.amount);
-      if (evaluated !== null) {
-        finalAmount = evaluated;
-      }
-    }
-
-    // Create updated values with evaluated amount for validation
-    const valuesToValidate = { ...values, amount: finalAmount };
-
-    if (!validateFields(valuesToValidate)) {
-      // Mirror Quick Add behavior: show a dialog with the validation error
-      const error = validateOperation(valuesToValidate, t);
-      if (error) {
-        showDialog(t('error'), error, [{ text: 'OK' }]);
-      }
-      return;
-    }
-
-    // Update the state with evaluated amount
-    setValues(v => ({ ...v, amount: finalAmount }));
-
-    // Pass the evaluated amount to prepareOperationData
-    const operationData = prepareOperationData(finalAmount);
-
-    if (isNew) {
-      await addOperation(operationData);
-    } else {
-      await updateOperation(operation.id, operationData);
-    }
-
-    // Save last accessed account
-    if (operationData.accountId) {
-      setLastAccessedAccount(operationData.accountId);
-    }
-
-    onClose();
-  }, [values, validateFields, prepareOperationData, isNew, addOperation, updateOperation, operation, onClose]);
-
-  const handleClose = useCallback(() => {
-    Keyboard.dismiss();
-    setErrors({});
-    onClose();
-  }, [onClose]);
-
-  // Check if operation belongs to a shadow category
-  const isShadowOperation = useMemo(() => {
-    if (!operation || !operation.categoryId) return false;
-    const category = categories.find(cat => cat.id === operation.categoryId);
-    return category?.isShadow || false;
-  }, [operation, categories]);
-
-  // Check if operation date is today (for shadow operations)
-  const isOperationToday = useMemo(() => {
-    if (!operation) return false;
-    const today = new Date().toISOString().split('T')[0];
-    return operation.date === today;
-  }, [operation]);
-
-  // Shadow operations can only be deleted if they were made today
-  const canDeleteShadowOperation = useMemo(() => {
-    return !isShadowOperation || isOperationToday;
-  }, [isShadowOperation, isOperationToday]);
-
-  const handleDelete = useCallback(() => {
-    if (onDelete && operation && canDeleteShadowOperation) {
-      onDelete(operation);
-      onClose();
-    }
-  }, [onDelete, operation, onClose, canDeleteShadowOperation]);
-
-  const openPicker = useCallback((type, data) => {
-    Keyboard.dismiss();
-    if (type === 'category') {
-      // For categories, show root folders and root entry categories
-      const rootItems = data.filter(cat => !cat.parentId);
-      setPickerState({ visible: true, type, data: rootItems, allCategories: data });
-      setCategoryNavigation({ currentFolderId: null, breadcrumb: [] });
-    } else {
-      setPickerState({ visible: true, type, data, allCategories: [] });
-    }
-  }, []);
-
-  const closePicker = useCallback(() => {
-    setPickerState({ visible: false, type: null, data: [], allCategories: [] });
-    setCategoryNavigation({ currentFolderId: null, breadcrumb: [] });
-  }, []);
-
-  const getAccountName = useCallback((accountId) => {
-    if (!accountId) return t('select_account');
-    const account = accounts.find(acc => acc.id === accountId);
-    return account ? account.name : t('select_account');
-  }, [accounts, t]);
-
-  const getCategoryName = useCallback((categoryId) => {
-    if (!categoryId) return t('select_category');
-    const displayName = getCategoryDisplayName(categoryId, categories, t);
-    return displayName || t('select_category');
-  }, [categories, t]);
-
-  // Navigate into a category folder
-  const navigateIntoFolder = useCallback((folder) => {
-    setPickerState(prev => {
-      const folderName = folder.nameKey ? t(folder.nameKey) : folder.name;
-      const children = prev.allCategories.filter(cat => cat.parentId === folder.id);
-
-      setCategoryNavigation(prevNav => ({
-        currentFolderId: folder.id,
-        breadcrumb: [...prevNav.breadcrumb, { id: folder.id, name: folderName }],
-      }));
-
-      return { ...prev, data: children };
-    });
-  }, [t]);
-
-  // Navigate back to previous folder level
-  const navigateBack = useCallback(() => {
-    setPickerState(prev => {
-      const newBreadcrumb = categoryNavigation.breadcrumb.slice(0, -1);
-      const newFolderId = newBreadcrumb.length > 0
-        ? newBreadcrumb[newBreadcrumb.length - 1].id
-        : null;
-
-      // Get the appropriate categories for this level
-      let newData;
-      if (newFolderId === null) {
-        newData = prev.allCategories.filter(cat => !cat.parentId);
-      } else {
-        newData = prev.allCategories.filter(cat => cat.parentId === newFolderId);
-      }
-
-      setCategoryNavigation({
-        currentFolderId: newFolderId,
-        breadcrumb: newBreadcrumb,
-      });
-
-      return { ...prev, data: newData };
-    });
-  }, [categoryNavigation.breadcrumb]);
-
-  const filteredCategories = useMemo(() => {
-    return categories.filter(cat => {
-      if (values.type === 'transfer') return false;
-      // Exclude shadow categories from selection
-      if (cat.isShadow) return false;
-      return cat.categoryType === values.type;
-    });
-  }, [categories, values.type]);
-
-  const formatDateForDisplay = useCallback((isoDate) => {
-    const date = new Date(isoDate);
-    return date.toLocaleDateString(undefined, {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-    });
-  }, []);
-
-  // Get source and destination accounts for currency detection
-  const sourceAccount = useMemo(() => {
-    return accounts.find(acc => acc.id === values.accountId);
-  }, [accounts, values.accountId]);
-
-  // Clear category when switching between expense and income to avoid mismatched categories
-  useEffect(() => {
-    // Only run when modal is visible (avoid interfering while modal closed)
-    if (!visible) return;
-
-    setValues(v => {
-      // If switching to transfer already handled elsewhere
-      if (v.type === 'transfer') return v;
-
-      // If current category exists but doesn't match new type, clear it
-      const cat = categories.find(c => c.id === v.categoryId);
-      if (v.categoryId && cat && cat.categoryType !== v.type) {
-        return { ...v, categoryId: '' };
-      }
-      return v;
-    });
-  }, [values.type, visible, categories]);
-
-  const destinationAccount = useMemo(() => {
-    return accounts.find(acc => acc.id === values.toAccountId);
-  }, [accounts, values.toAccountId]);
-
-  // Check if this is a multi-currency transfer
-  const isMultiCurrencyTransfer = useMemo(() => {
-    if (values.type !== 'transfer') return false;
-    if (!sourceAccount || !destinationAccount) return false;
-    return sourceAccount.currency !== destinationAccount.currency;
-  }, [values.type, sourceAccount, destinationAccount]);
-
-  // Auto-populate exchange rate when accounts change
-  useEffect(() => {
-    if (isMultiCurrencyTransfer && sourceAccount && destinationAccount && !values.exchangeRate) {
-      const rate = Currency.getExchangeRate(sourceAccount.currency, destinationAccount.currency);
-      if (rate) {
-        setValues(v => ({ ...v, exchangeRate: rate }));
-        setLastEditedField('exchangeRate');
-      }
-    }
-  }, [isMultiCurrencyTransfer, sourceAccount, destinationAccount, values.exchangeRate]);
-
-  // Auto-calculate based on which field was last edited
-  useEffect(() => {
-    if (!isMultiCurrencyTransfer) {
-      // Clear exchange rate fields for same-currency transfers
-      if (values.exchangeRate || values.destinationAmount) {
-        setValues(v => ({ ...v, exchangeRate: '', destinationAmount: '' }));
-        setLastEditedField(null);
-      }
-      return;
-    }
-
-    if (!sourceAccount || !destinationAccount) return;
-
-    // If user edited destination amount, calculate the rate
-    if (lastEditedField === 'destinationAmount') {
-      if (values.amount && values.destinationAmount) {
-        // Calculate rate = destinationAmount / sourceAmount (accounting for decimal places)
-        const sourceAmount = parseFloat(values.amount);
-        const destAmount = parseFloat(values.destinationAmount);
-
-        if (!isNaN(sourceAmount) && !isNaN(destAmount) && sourceAmount > 0) {
-          const calculatedRate = (destAmount / sourceAmount).toFixed(6);
-          // Only update if the rate has meaningfully changed (avoid precision issues)
-          const currentRate = parseFloat(values.exchangeRate || '0');
-          const newRate = parseFloat(calculatedRate);
-          if (Math.abs(currentRate - newRate) > 0.000001) {
-            setValues(v => ({ ...v, exchangeRate: calculatedRate }));
-          }
-        }
-      }
-    }
-    // If user edited amount or rate, calculate destination amount
-    else if (lastEditedField === 'amount' || lastEditedField === 'exchangeRate') {
-      if (values.amount && values.exchangeRate) {
-        const converted = Currency.convertAmount(
-          values.amount,
-          sourceAccount.currency,
-          destinationAccount.currency,
-          values.exchangeRate,
-        );
-        if (converted && converted !== values.destinationAmount) {
-          setValues(v => ({ ...v, destinationAmount: converted }));
-        }
-      }
-    }
-  }, [isMultiCurrencyTransfer, values.amount, values.exchangeRate, values.destinationAmount, sourceAccount, destinationAccount, lastEditedField]);
-
-  // Prepare operation data with currency information for saving
-  const prepareOperationData = useCallback((customAmount = null) => {
-    const data = { ...values };
-
-    // Override amount if provided (used when evaluating expressions)
-    if (customAmount !== null) {
-      data.amount = customAmount;
-    }
-
-    if (isMultiCurrencyTransfer && sourceAccount && destinationAccount) {
-      data.sourceCurrency = sourceAccount.currency;
-      data.destinationCurrency = destinationAccount.currency;
-    }
-
-    // Ensure amount is preserved when editing
-    if (!isNew && !data.amount && operation?.amount) {
-      data.amount = String(operation.amount);
-    }
-
-    // Format amounts based on currency decimal places
-    if (data.amount && sourceAccount) {
-      data.amount = Currency.formatAmount(data.amount, sourceAccount.currency);
-    }
-
-    if (data.destinationAmount && destinationAccount) {
-      data.destinationAmount = Currency.formatAmount(data.destinationAmount, destinationAccount.currency);
-    }
-
-    return data;
-  }, [values, isMultiCurrencyTransfer, sourceAccount, destinationAccount, isNew, operation]);
+  }, [isShadowOperation, setValues, setLastEditedField]);
 
   const TYPES = [
     { key: 'expense', label: t('expense'), icon: 'minus-circle' },
